fix: find instead of findone and doc toArray

+++ cmds/resetConfig.js
@@ -111,11 +111,11 @@ module.exports.run = async (bot, message, args, guild) => {
   const discordId = targetUser.id;
 
   statusmsg = await message.channel.send(`:thinking: Accessing database...`);
-  let doc = await mongoDB().collection("users").findOne({ discordId });
+  let doc = await mongoDB().collection("users").find({ discordId }).toArray();
   if(doc) {
-    if (doc.docs.length === 1) {
+    if (doc.length === 1) {
       await statusmsg.edit(`:thinking: User Found...`);
-      const uid = doc.docs[0].id;
+      const uid = doc.docs[0].uid;
       await statusmsg.edit(`:thinking: Resetting config...`);
       await mongoDB()
         .collection("configs")
@@ -125,7 +125,7 @@ module.exports.run = async (bot, message, args, guild) => {
         message: `:white_check_mark: Done`,
       };
     }
-    else {
+    else if (doc.length > 1) {
       await statusmsg.delete();
       return {
         status: true,
@@ -143,6 +143,6 @@ module.exports.run = async (bot, message, args, guild) => {
 };
 
 module.exports.cmd = {
-  name: "resetConfig",
+  name: "resetconfig",
   needMod: true,
 };
